refactor: Moved ForgotPwd module
