+++ app/src/main/res/layout/activity_forgot_pwd.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_margin="32dp"
     android:orientation="vertical"
-    tools:context=".ui.forgotpwd.ForgotPwdActivity">
+    tools:context=".ui.sign.ForgotPwdActivity">
 
     <!-- TODO Implement -->
